Se hicieron ajustes

+++ Recetario/app/src/main/java/com/capi/recetario/Ingrediente.kt
@@ -2,10 +2,10 @@ package com.capi.recetario
 
 class Ingrediente {
     var nombreIngrediente: String
-    var cantidadIngrediente: Int
+    var cantidadIngrediente: Double
     var tipoIngrediente: String
 
-    constructor(nombreIngrediente: String, cantidadIngrediente: Int, tipoIngrediente: String){
+    constructor(nombreIngrediente: String, cantidadIngrediente: Double, tipoIngrediente: String){
         this.nombreIngrediente = nombreIngrediente
         this.cantidadIngrediente = cantidadIngrediente
         this.tipoIngrediente = tipoIngrediente
